Switch to override (#887)
